Add skeleton 'init' and 'verify' commands to CLI.
