+++ sameproject/cli/__init__.py
@@ -1,2 +1,4 @@
 from .version import version
+from .verify import verify
+from .init import init
 from .run import run
