+++ sameproject/cli/init.py
@@ -1,0 +1,9 @@
+import click
+
+
+@click.command()
+def init():
+    """Creates a new SAME config file."""
+    print("The init command has not been implemented yet.")
+    exit(1)
+
