+++ sameproject/cli/verify.py
@@ -1,0 +1,8 @@
+import click
+
+
+@click.command()
+def verify():
+    """Verifies a SAME config file."""
+    print("The verify command has not been implemented yet.")
+    exit(1)
