+++ test/cli/test_compile.py
@@ -1,7 +1,7 @@
 from sameproject.ops.backends import render as template_render
 from sameproject.ops import notebooks
 from sameproject.ops.backends import deploy
-from sameproject.data.same_config import SameConfig
+from sameproject.data.config import SameConfig
 from click.testing import CliRunner
 from sameproject.cli import run
 from pathlib import Path
