+++ test/sdk/test_dataset.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 sys.path.append(Path(__file__).parent.parent.absolute().name)
 import sameproject.sdk as same
-from sameproject.data.same_config import SameConfig
+from sameproject.data.config import SameConfig
 import os
 import pytest
 from pytest_mock import mocker
